+++ config/index.js
@@ -1,1 +1,3 @@
 export const API_URL = process.env.NEXT_PUBLIC_API_URL || 'http://localhost:1337'
+
+export const PER_PAGE = 5
