+++ pages/events/index.js
@@ -1,11 +1,12 @@
 import Layout from '@/components/Layout'
+import Pagination from '@/components/Pagination'
 import EventItem from '@/components/EventItem'
-import { API_URL } from '@/config/index'
-import Link from 'next/link'
-const PER_PAGE = 2
+import { API_URL, PER_PAGE } from '@/config/index'
+
+
 
 export default function EventsPage({ events, page, total }) {
-  const lastPage = Math.ceil(total / PER_PAGE)
+
 
   return (
     <Layout>
@@ -15,17 +16,7 @@ export default function EventsPage({ events, page, total }) {
         <EventItem key={evt.id} evt={evt} />
       ))}
 
-      {page > 1 && (
-        <Link href={`/events?page=${page - 1}`}>
-          <a className='btn-secondary'>Prev</a>
-        </Link>
-      )}
-
-      {page < lastPage && (
-        <Link href={`/events?page=${page + 1}`}>
-          <a className='btn-secondary'>Next</a>
-        </Link>
-      )}
+      <Pagination page={page} total={total} />
     </Layout>
   );
 }
